Update unfair table
Hey guys, I love what you do, but this table contains some bullshit :)

> Data can be viewed in a normal Web browser
Sorry guys, but that's a stupid question. Every decentralized storage solution can be viewable in browser with a CDN (or middle server). Everything that is really decentralized needs a full node. Because of that, it's not viewable in a browser, without a full node. Gaia will be decentralized when I can add multiple storages to it, but that's optional.

> Data can be deleted
On Sia, files can be deleted, I think this is just a simple mistake.

> Data lookups have predictable performance
I'm watching 1080p films from a mounted Sia drive. That's improved a lot and will be improved more.

> Supports multiple backends natively
Ahh, here we go again. What does this question mean? On Sia, every file stored on 30 geographically distributed hosts. I think it is multiple. You can increase/decrease the redundancy also.
I removed this line, because every competitor is a backend, while Gaia is the client. You can use Sia with Gaia (with a mounted s3 drive), that's a stupid comparison. Apple is not comparable with pear.

> Data is read/write
What does this question mean?

Note that I have strong experience with Sia, but I don't know much about Dat and SSB (that's a social media, not storage, why is it on the list? Apple and pear again.) IPFS is a protocol, FileCoin is the storage (yeah, not working yet, but again, that's not compareable).